requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# c:\Mali\Hobby\Udacity\CT-Recon\CT-BackProjection.py: 45
Pillow == 7.2.0

# c:\Mali\Hobby\Udacity\CT-Recon\CT-BackProjection.py: 46
matplotlib == 3.3.2

# c:\Mali\Hobby\Udacity\CT-Recon\CT-BackProjection.py: 47
numpy == 1.19.2

# c:\Mali\Hobby\Udacity\CT-Recon\CT-BackProjection.py: 48
scipy == 1.5.2
